generate search links from patterns provided by the config
The new search_links config option consists of a list of maps, each containing a pattern and a
template. The pattern is matched against the full URL of each repo. If a pattern matches, the
corresponding template is evaluated and used as the search link. The template associated with
the first matching pattern is used. If no patterns match, the unmodified repo URL is used.